# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DumbPad API",
    "version": "1.0.8",
    "description": "DumbPad content-management API. Legacy deployments authenticate with a PIN; AUTH_V2_ENABLED deployments use an opaque browser session or a scoped API token. API tokens cannot access authentication or data-management routes."
  },
  "servers": [{ "url": "/" }],
  "security": [{ "cookieAuth": [] }, { "sessionCookieAuth": [] }, { "apiTokenBearer": [] }, { "pinBearer": [] }],
  "paths": {
...
    "/api/assets/images": {
      "post": {
        "summary": "Upload an image asset while preserving its original bytes",
        "parameters": [{ "name": "X-Asset-Name", "in": "header", "description": "URL-encoded filename for display and download", "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "image/jpeg": { "schema": { "type": "string", "format": "binary" } }, "image/png": { "schema": { "type": "string", "format": "binary" } }, "image/webp": { "schema": { "type": "string", "format": "binary" } }, "image/avif": { "schema": { "type": "string", "format": "binary" } }, "image/gif": { "schema": { "type": "string", "format": "binary" } } } },
        "responses": { "201": { "description": "Created image asset", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ImageAsset" } } } }, "413": { "$ref": "#/components/responses/Error" }, "415": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/assets/files": {
      "post": {
        "summary": "Upload an allow-listed ordinary file asset for article download",
        "parameters": [
          { "name": "X-Asset-Name", "in": "header", "required": true, "description": "URL-encoded filename", "schema": { "type": "string" } },
          { "name": "X-Asset-Type", "in": "header", "description": "Browser-reported MIME, checked against the filename extension", "schema": { "type": "string" } }
        ],
        "requestBody": { "required": true, "content": { "application/octet-stream": { "schema": { "type": "string", "format": "binary" } } } },
        "responses": { "201": { "description": "Created ordinary file asset", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FileAsset" } } } }, "413": { "$ref": "#/components/responses/Error" }, "415": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/assets/{id}/{variant}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^[a-fA-F0-9-]{16,64}$" } },
        { "name": "variant", "in": "path", "required": true, "schema": { "type": "string", "enum": ["preview", "original", "download"] } }
      ],
      "get": { "summary": "Read an image preview, original, or ordinary-file download", "responses": { "200": { "description": "Binary asset response" }, "404": { "$ref": "#/components/responses/Error" } } }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": { "type": "apiKey", "in": "cookie", "name": "dumbpad_auth" },
      "sessionCookieAuth": { "type": "apiKey", "in": "cookie", "name": "dumbpad_auth_session" },
      "apiTokenBearer": { "type": "http", "scheme": "bearer", "description": "Scoped V2 content API token; cannot access auth or data-management routes." },
      "pinBearer": { "type": "http", "scheme": "bearer", "description": "Legacy DumbPad PIN for trusted automation only. Disabled in V2." }
    },
...
    "responses": { "Error": { "description": "Request failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } } },
    "schemas": {
      "Error": { "type": "object", "properties": { "error": { "type": "string" }, "currentVersion": { "type": "integer" } }, "required": ["error"] },
...
      "ImageAsset": { "type": "object", "required": ["id", "assetId", "name", "type", "size", "kind", "previewUrl", "originalUrl", "downloadUrl"], "properties": { "id": { "type": "string" }, "assetId": { "type": "string" }, "name": { "type": "string" }, "type": { "type": "string", "pattern": "^image/" }, "kind": { "const": "image" }, "size": { "type": "integer", "format": "int64" }, "previewUrl": { "type": "string" }, "originalUrl": { "type": "string" }, "downloadUrl": { "type": "string" } } },
      "FileAsset": { "type": "object", "required": ["id", "assetId", "name", "type", "size", "kind", "previewUrl", "originalUrl", "downloadUrl"], "properties": { "id": { "type": "string" }, "assetId": { "type": "string" }, "name": { "type": "string" }, "type": { "type": "string" }, "kind": { "const": "file" }, "size": { "type": "integer", "format": "int64", "maximum": 20971520 }, "previewUrl": { "type": "null" }, "originalUrl": { "type": "string" }, "downloadUrl": { "type": "string" } } },
...
    }
  }
}